Написал объект для film-details, осталось добавить в него пару значений
