+++ src/mock/const.js
@@ -1,0 +1,91 @@
+const FILM_NAMES = [
+  `Привет`,
+  `Новый фильм`,
+  `2012`,
+  `Ещё что-то`,
+  `Почти пятнадцатый`,
+  `Супер мок`,
+  `Намного проще`,
+  `Интенсив`,
+  `Конец зимы`,
+  `Идём на React`,
+  `Node JS`,
+  `Ещё какой-то фильм`,
+  `Первый`,
+  `Второй`,
+  `Десятый`
+];
+
+const GENRES = [
+  `Musical`,
+  `Western`,
+  `Comedy`,
+  `Cartoon`,
+  `Mystery`
+];
+
+const POSTERS = [
+  `made-for-each-other.png`,
+  `popeye-meets-sinbad.png`,
+  `sagebrush-trail.jpg`,
+  `santa-claus-conquers-the-martians.jpg`,
+  `the-dance-of-life.jpg`,
+  `the-great-flamarion.jpg`,
+  `the-man-with-the-golden-arm.jpg`
+];
+
+const DESCRIPTION = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`;
+
+const REGGISEURS = [
+  `Martin`,
+  `Max`,
+  `Adam`,
+  `Alex`,
+  `John`,
+  `July`,
+  `Someby`,
+  `Anybody`,
+  `Nobody`
+];
+
+const REGGISEURS = [
+  `Martin`,
+  `Max`,
+  `Adam`,
+  `Alex`,
+  `John`,
+  `July`,
+  `Someby`,
+  `Anybody`,
+  `Nobody`
+];
+
+const SCENARISTS = [
+  `Lina`,
+  `Paul`,
+  `Nick`,
+  `Lisa`,
+  `Ivan`,
+  `Maria`
+];
+
+const ACTORS = [
+  `Anna`,
+  `David`,
+  `Serg`,
+  `Eva`,
+  `Malh`,
+  `Laos`
+];
+
+const COUNTRIES = [
+  `Russia`,
+  `Germany`,
+  `USA`,
+  `China`,
+  `Burkino Faso`,
+  `Ukraine`,
+  `Austria`
+];
+
+export {FILM_NAMES, GENRES, POSTERS, DESCRIPTION, REGGISEURS, SCENARISTS, ACTORS, COUNTRIES};
