+++ src/util.js
@@ -1,0 +1,17 @@
+const getRandomIntegerNumber = (min, max) => {
+  return min + Math.floor(max * Math.random());
+};
+
+const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomIntegerNumber(0, array.length);
+
+  return array[randomIndex];
+};
+
+const getRandomDescription = (string) => {
+  return string.split(`.`)
+    .filter(() => Math.random() > 0.5)
+    .slice(0, getRandomIntegerNumber(1, 3));
+};
+
+export {getRandomArrayItem, getRandomIntegerNumber, getRandomDescription};
